+++ src/main/gen/com/dvd/intellij/d2/lang/_D2Lexer.java
@@ -5,9 +5,9 @@ package com.dvd.intellij.d2.lang;
 import com.intellij.lexer.FlexLexer;
 import com.intellij.psi.tree.IElementType;
 
+import static com.dvd.intellij.d2.lang.D2ElementTypes.*;
 import static com.intellij.psi.TokenType.BAD_CHARACTER;
 import static com.intellij.psi.TokenType.WHITE_SPACE;
-import static com.dvd.intellij.d2.lang.D2ElementTypes.*;
 
 
 /**