+++ src/main/kotlin/com/dvd/intellij/d2/components/D2Layout.kt
@@ -20,7 +20,5 @@ data class D2Layout(
     return true
   }
 
-  override fun hashCode(): Int {
-    return name.hashCode()
-  }
+  override fun hashCode(): Int = name.hashCode()
 }