+++ CHANGELOG.md
@@ -3,5 +3,15 @@
 # d2-intellij Changelog
 
 ## [Unreleased]
+
 ### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+
+- Completion contributor
+- Formatter
+
+## [0.0.1-eap]
+
+### Added
+
+- Initial scaffold created
+  from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)